requirements.txt
torch>=2.0
torchvision
transformers>=4.35
streamlit>=1.30
plotly>=5.18
matplotlib>=3.8
opencv-python>=4.8
numpy>=1.24
scipy>=1.11
Pillow>=10.0
facenet-pytorch>=2.5
scikit-learn>=1.3
grad-cam>=1.4
PyWavelets>=1.4
